+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Created `setTime`, `setMilliseconds`, `setSeconds`, `setMinutes`, `setHours` functions
+
 ## 0.5.0
 - Created **start** and **end** functions